+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.15.2](https://github.com/Kong/spec-renderer/compare/v1.15.1...v1.15.2) (2024-05-31)
+
+
+### Bug Fixes
+
+* **sandbox:** vite config ([#100](https://github.com/Kong/spec-renderer/issues/100)) ([b379bad](https://github.com/Kong/spec-renderer/commit/b379bad1cc25ce9aba5b53612d496b8db2107fd0))
+
+## [1.15.1](https://github.com/Kong/spec-renderer/compare/v1.15.0...v1.15.1) (2024-05-31)
+
+
+### Bug Fixes
+
+* **wasm:** enable ssr wasm support [KHCP-12146] ([#97](https://github.com/Kong/spec-renderer/issues/97)) ([1049d5f](https://github.com/Kong/spec-renderer/commit/1049d5f59b4b77c1b1185626f045ee62db715a3e))
+
+# [1.15.0](https://github.com/Kong/spec-renderer/compare/v1.14.1...v1.15.0) (2024-05-31)
+
+
+### Features
+
+* schema section base styling [KHCP-12131, KHCP-12130] ([#96](https://github.com/Kong/spec-renderer/issues/96)) ([79246cc](https://github.com/Kong/spec-renderer/commit/79246cc376d4da630b6a95deb9e42dc467610b19))
+
 ## [1.14.1](https://github.com/Kong/spec-renderer/compare/v1.14.0...v1.14.1) (2024-05-31)